requirements.txt.txt
PyOpenGL>=3.1.7
PyOpenGL_accelerate>=3.1.7
glfw>=2.7.0
